% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{auto2}
\alias{auto2}
\title{1978 Automobile Data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 74 rows and 12 columns.}
\source{
\url{http://www.stata-press.com/data/r14/g.html}
}
\usage{
auto2
}
\description{
From Consumer Reports with permission. Same as \link[sysuse:auto]{auto.dta}, except repair value label is added to rep78.
}
\keyword{datasets}
