% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{autornd}
\alias{autornd}
\title{1978 Automobile Data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 74 rows and 3 columns.}
\source{
\url{http://www.stata-press.com/data/r14/g.html}
}
\usage{
autornd
}
\description{
Produced from auto.dta mpg rounded to 5 mpg increments 3 weight rounded to 500 lb increments.
}
\keyword{datasets}
