% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cancer}
\alias{cancer}
\title{cancer}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 48 rows and 8 columns.}
\source{
\url{http://www.stata-press.com/data/r14/g.html}
}
\usage{
cancer
}
\description{
Patient Survival in Drug Trial.
}
\keyword{datasets}
