% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census}
\alias{census}
\title{census}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 50 rows and 13 columns.}
\source{
\url{http://www.stata-press.com/data/r14/g.html}
}
\usage{
census
}
\description{
1980 Census data by state
}
\keyword{datasets}
