% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{citytemp}
\alias{citytemp}
\title{citytemp}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 956 rows and 6 columns.}
\source{
\url{http://www.stata-press.com/data/r14/g.html}
}
\usage{
citytemp
}
\description{
City Temperature Data
}
\keyword{datasets}
