% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{educ99gdp}
\alias{educ99gdp}
\title{Education and GDP}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 3 columns.}
\source{
\url{http://www.stata-press.com/data/r14/g.html}
}
\usage{
educ99gdp
}
\description{
Spending on tertiary education as % of GDP, 1999 2 Source: OECD, Education at a Glance, 2002.
}
\keyword{datasets}
