% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gnp96}
\alias{gnp96}
\title{gnp96}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 142 rows and 2 columns.}
\source{
\url{http://www.stata-press.com/data/r14/g.html}
}
\usage{
gnp96
}
\description{
Source: U.S. Department of Commerce, Bureau of Economic Analysis.
}
\keyword{datasets}
