% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lifeexp}
\alias{lifeexp}
\title{Life expectancy, 1998}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 68 rows and 6 columns.}
\source{
\url{http://www.stata-press.com/data/r14/g.html}
}
\usage{
lifeexp
}
\description{
Source: The World Bank Group, Learning Modules, http://www.worldbank.org/depweb/english/modules/basdata/bdata.
}
\keyword{datasets}
