% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{network1a}
\alias{network1a}
\title{network1a}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 77 rows and 11 columns.}
\source{
\url{http://www.stata-press.com/data/r14/g.html}
}
\usage{
network1a
}
\description{
To be added.
}
\keyword{datasets}
