% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nlsw88}
\alias{nlsw88}
\title{National Longitudinal Survey of Young Woman}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2246 rows and 17 columns.}
\source{
\url{http://www.stata-press.com/data/r14/g.html}
}
\usage{
nlsw88
}
\description{
1988 data, extracted from National Longitudinal of Young Woman who were ages 14-24 in 1968 (NLSW). This dataset is the result of extraction and processing by various people at various times. For more information on the NLS, see http://www.bls.gov/nls/.
}
\keyword{datasets}
