% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nlswide1}
\alias{nlswide1}
\title{National Longitudinal Survey of Young Woman}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9 rows and 19 columns.}
\source{
\url{http://www.stata-press.com/data/r14/g.html}
}
\usage{
nlswide1
}
\description{
Young Women 14-26 years of age in 1968
}
\keyword{datasets}
