% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop2000}
\alias{pop2000}
\title{pop2000}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 17 rows and 20 columns.}
\source{
\url{http://www.stata-press.com/data/r14/g.html}
}
\usage{
pop2000
}
\description{
Source: U.S. Census Bureau, Census 2000, PHC-T-9, Tables 1-3: Table 1. Total Pop. by Age, Race and Hispanic or Latino Origin; Table 2. Male Pop. by Age, Race and Hispanic or Latino Origin; Table 3. Female Pop. by Age, Race and Hispanic or Latino Origin.
}
\keyword{datasets}
