% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sandstone}
\alias{sandstone}
\title{Subsea elevation of Lamont sandstone in an area of Ohio}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6400 rows and 4 columns.}
\source{
\url{http://www.stata-press.com/data/r14/g.html}
}
\usage{
sandstone
}
\description{
Data is artificial data based on a dataset from Statistics and Data Analysis in Geology, 3rd edition, by John C Davis.
}
\keyword{datasets}
