% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp500}
\alias{sp500}
\title{S&P 500}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 248 rows and 7 columns.}
\source{
\url{http://www.stata-press.com/data/r14/g.html}
}
\usage{
sp500
}
\description{
S&P 500 Historical Data. Source: Yahoo!Finance, Yahoo! Inc., and Commodity Systems, Inc.
}
\keyword{datasets}
