% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tsline1}
\alias{tsline1}
\title{tsline1}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 200 rows and 3 columns.}
\source{
\url{http://www.stata-press.com/data/r14/g.html}
}
\usage{
tsline1
}
\description{
To be added.
}
\keyword{datasets}
