% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tsline2}
\alias{tsline2}
\title{tsline2}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 365 rows and 4 columns.}
\source{
\url{http://www.stata-press.com/data/r14/g.html}
}
\usage{
tsline2
}
\description{
To be added.
}
\keyword{datasets}
