% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uslifeexp}
\alias{uslifeexp}
\title{U.S. life expectancy, 1900-1999}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 100 rows and 10 columns.}
\source{
\url{http://www.stata-press.com/data/r14/g.html}
}
\usage{
uslifeexp
}
\description{
For selected years, life table values shown are estimates. Table 12. Estimated life expectancy at birth in years, by race and sex. Source: https://www.cdc.gov/nchs/data/nvsr/nvsr50/nvsr50_06.pdf National Vital Statistics Report, Vol. 50 No. 6, 21mar2002
}
\keyword{datasets}
