% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uslifeexp2}
\alias{uslifeexp2}
\title{U.S. life expectancy, 1900-1940}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 41 rows and 2 columns.}
\source{
\url{http://www.stata-press.com/data/r14/g.html}
}
\usage{
uslifeexp2
}
\description{
This dataset is a subset of \link[sysuse:uslifeexp]{uslifeexp.dta}. For selected years, life table values shown are estimates. Table 12. Estimated life expectancy at birth in years, by race and sex. Source: https://www.cdc.gov/nchs/data/nvsr/nvsr50/nvsr50_06.pdf, National Vital Statistics Report, Vol. 50 No. 6, 21mar2002.
}
\keyword{datasets}
