% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{voter}
\alias{voter}
\title{voter}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15 rows and 5 columns.}
\source{
\url{http://www.stata-press.com/data/r14/g.html}
}
\usage{
voter
}
\description{
To be added.
}
\keyword{datasets}
